Internal change
PiperOrigin-RevId: 267007907
